typo(dubbo-monitor): Fix typo of DubboMonitorFactory
