+++ dubbo-monitor/dubbo-monitor-default/src/main/resources/META-INF/dubbo/internal/com.alibaba.dubbo.monitor.MonitorFactory
@@ -1,1 +1,1 @@
-dubbo=com.alibaba.dubbo.monitor.dubbo.DubboMonitorFactroy
+dubbo=com.alibaba.dubbo.monitor.dubbo.DubboMonitorFactory
